Brain-Tumor-Detector/back-end/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pillow
numpy